requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
django-payments-paybox==0.2.3
django-perms==2.1.0
django-phonenumber-field==5.2.0
django-rest-framework==0.1.0
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
logger==1.4
magic-filter==1.0.9
MarkupSafe==2.1.1
multidict==6.0.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==21.3
phonenumbers==8.13.0
Pillow==9.3.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pyparsing==3.0.9
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.6
PyYAML==5.4.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
telebot==0.0.4
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
yarl==1.8.2
